Add folder support

+++ player.py
@@ -1,6 +1,7 @@
 import pathlib
 import random
 import subprocess
+from sys import stderr
 from typing import List, Union
 import os
 import argparse
@@ -8,37 +9,61 @@ import logging
 import config
 
 
-def isSonata(f: str) -> bool:
-    return len(f) > 7 and f[-7] == ' ' and f[-6].isdigit()
-
-
 class player:
     """ Playing the music in a directory (and its subdirectories)
     """
 
-    def __init__(self, path) -> None:
+    def __init__(self) -> None:
         self.keepplay = True
         self.P = None
         self.next = []
-
         self.musics = []
+        
+    def set_musics(self, paths: List[str]) -> None:
+        """ Get musics under paths
+        """
+        self.musics = []
+        for p in paths:
+            self.musics.extend(self._get_one_path(p))
+
+    def _get_one_path(self, path: str) -> List[List[str]]:
+        """ Fetch musics under path.
+
+        Args:
+            path (str): path containing music files
+
+        Returns:
+            List[List[str]]: all musics under path, each element in the 
+            result represent:
+                1. Single file music: [file]
+                2. Sonata: [[part_last, ..., part_2, part_1]] 
+        """
+        res = []
         tmp = {}
+
         for dirpath, _, filenames in os.walk(path):
             for name in filenames:
                 if name.endswith(".flac"):
                     f = os.path.join(dirpath, name)
-                    if not isSonata(f):
-                        self.musics.append([f])
+                    if not player.isSonata(f):
+                        res.append([f])
                     else:
                         tmp.setdefault(f[0:-7], []).append(f)
-        for key in tmp:
-            self.musics.append(sorted(tmp[key]))
+        for m in tmp.values():
+            res.append(sorted(m, reverse=True))
+        
+        return res
+
+    @staticmethod
+    def isSonata(f: str) -> bool:
+        return len(f) > 7 and f[-7] == ' ' and f[-6].isdigit()
+
 
     def start(self):
         """ Blockingly play musics
         """
         if 0 == len(self.musics):
-            raise "Did not find any music file under this folder."
+            raise Exception("Did not find any music file under this folder.")
 
         if self.isplaying():
             return
@@ -48,7 +73,10 @@ class player:
             m = self.NextSong()
             logging.info(m)
             self.P = subprocess.Popen(
-                ["ffplay", "-hide_banner", "-autoexit", "-nodisp", m])
+                ["ffplay", "-hide_banner", "-autoexit", "-nodisp", m], 
+                # stdout=subprocess.DEVNULL,
+                # stderr = subprocess.DEVNULL
+            )
             self.P.wait()
 
     def NextSong(self) -> str:
@@ -61,9 +89,10 @@ class player:
     def stop(self):
         self.keepplay = False
         self.next = []
-        self.P.kill()
-        self.P.wait()
-        self.P = None
+        if self.isplaying():
+            self.P.kill()
+            self.P.wait()
+            self.P = None
 
     def isplaying(self):
         return self.P != None
@@ -72,9 +101,10 @@ class player:
 if __name__ == "__main__":
 
     parser = argparse.ArgumentParser(description="Simple music player.")
-    parser.add_argument("path", metavar="path", nargs='?', default=config.DefaultPath,
-                        help="Path to music folder, default to system environment variable")
-    arguments = parser.parse_args()
-    p = player(arguments.path)
+    parser.add_argument("paths", action = 'append',
+                        help="Path(s) to music folder")
+    args = parser.parse_args()
+    p = player()
+    p.set_musics(args.paths)
 
     p.start()
